api/requirements.txt
aerich
asyncpg
fastapi
tortoise-orm
uvicorn
python-dotenv
passlib
python-jose
bcrypt